backend/app/requirements.txt
fastapi[all]==0.92.0
pydantic==1.10.5
SQLAlchemy==2.0.2
sqlalchemy-utils
sqlalchemy_mixins
sqlalchemy_views==0.3.2
sqlalchemy-stubs
alembic==1.9.4
uvicorn
gunicorn
pydantic[email]==1.10.5
passlib[bcrypt]
tenacity==8.2.1
python-multipart~=0.0.5
asyncpg~=0.27.0
asyncpg_utils~=0.6.0
psycopg2-binary
psycopg2
python-jose~=3.3.0
pytest~=7.2.0
bcrypt~=4.0.1
httpx~=0.23.0
python-dotenv
loguru==0.6.0
redis==4.5.1
celery
raven
pyamqp
pytz


faker==17.0.0
pytest
pytest-asyncio
black==23.1.0
isort
flake8
mypy==1.0.1
